[feat]: [PL-25546]: AccessControl check for SMTP (#34530)
* [PL-25546]: Adding ACS to SMTP

* [PL-25546]: Adding ResourceScope

* [PL-25546]: Added SMTP Configs to yml

* [PL-25546]: Added SMTP Configs to yml #2

* [PL-25546]: Added ResourceType

* [PL-25546]: Version Updated

* [PL-25546]: Testcase added

* [PL-25546]: Testcase added #2

* [PL-25546]: Review Comment Addressed

* [PL-25546]: Fixed Version

* [PL-25546]: Removed Test

* [PL-25546]: Resolved Unit Test

* [PL-25546]: Test Names

+++ 120-ng-manager/src/main/java/io/harness/ng/core/smtp/resources/SmtpNgResource.java
@@ -9,12 +9,19 @@ package io.harness.ng.core.smtp.resources;
 
 import static io.harness.NGCommonEntityConstants.ACCOUNT_PARAM_MESSAGE;
 import static io.harness.annotations.dev.HarnessTeam.PL;
+import static io.harness.connector.accesscontrol.SMTPAccessControlPermissions.DELETE_SMTP_PERMISSION;
+import static io.harness.connector.accesscontrol.SMTPAccessControlPermissions.EDIT_SMTP_PERMISSION;
+import static io.harness.connector.accesscontrol.SMTPAccessControlPermissions.VIEW_SMTP_PERMISSION;
 
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 
 import io.harness.NGCommonEntityConstants;
 import io.harness.accesscontrol.AccountIdentifier;
+import io.harness.accesscontrol.acl.api.Resource;
+import io.harness.accesscontrol.acl.api.ResourceScope;
+import io.harness.accesscontrol.clients.AccessControlClient;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.connector.accesscontrol.ResourceTypes;
 import io.harness.ng.core.dto.ErrorDTO;
 import io.harness.ng.core.dto.FailureDTO;
 import io.harness.ng.core.dto.NgSmtpDTO;
@@ -77,7 +84,14 @@ import lombok.extern.slf4j.Slf4j;
     })
 @OwnedBy(PL)
 public class SmtpNgResource {
-  @Inject private SmtpNgService smtpNgService;
+  private final SmtpNgService smtpNgService;
+  private final AccessControlClient accessControlClient;
+
+  @Inject
+  public SmtpNgResource(SmtpNgService smtpNgService, AccessControlClient accessControlClient) {
+    this.smtpNgService = smtpNgService;
+    this.accessControlClient = accessControlClient;
+  }
 
   @POST
   @ApiOperation(value = "Create SMTP config", nickname = "createSmtpConfig")
@@ -93,6 +107,8 @@ public class SmtpNgResource {
   save(@Valid @NotNull NgSmtpDTO variable,
       @Parameter(description = ACCOUNT_PARAM_MESSAGE, required = true) @NotNull @AccountIdentifier @QueryParam(
           NGCommonEntityConstants.ACCOUNT_KEY) String accountIdentifier) throws IOException {
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, null, null), Resource.of(ResourceTypes.SMTP, null), EDIT_SMTP_PERMISSION);
     variable.setAccountId(accountIdentifier);
     NgSmtpDTO response = smtpNgService.saveSmtpSettings(variable);
     return ResponseDTO.newResponse(response);
@@ -131,6 +147,8 @@ public class SmtpNgResource {
   update(@Valid @NotNull NgSmtpDTO variable,
       @Parameter(description = ACCOUNT_PARAM_MESSAGE, required = true) @NotNull @AccountIdentifier @QueryParam(
           NGCommonEntityConstants.ACCOUNT_KEY) String accountIdentifier) throws IOException {
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, null, null), Resource.of(ResourceTypes.SMTP, null), EDIT_SMTP_PERMISSION);
     variable.setAccountId(accountIdentifier);
     NgSmtpDTO response = smtpNgService.updateSmtpSettings(variable);
     return ResponseDTO.newResponse(response);
@@ -174,6 +192,8 @@ public class SmtpNgResource {
   delete(@Parameter(description = "Config identifier") @PathParam("identifier") String identifier,
       @Parameter(description = ACCOUNT_PARAM_MESSAGE, required = true) @NotNull @AccountIdentifier @QueryParam(
           NGCommonEntityConstants.ACCOUNT_KEY) String accountIdentifier) throws IOException {
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountIdentifier, null, null), Resource.of(ResourceTypes.SMTP, null), DELETE_SMTP_PERMISSION);
     Boolean response = smtpNgService.deleteSmtpSettings(identifier);
     return ResponseDTO.newResponse(response);
   }
@@ -191,6 +211,8 @@ public class SmtpNgResource {
   @ExceptionMetered
   public ResponseDTO<NgSmtpDTO>
   get(@Parameter(description = ACCOUNT_PARAM_MESSAGE) @QueryParam("accountId") String accountId) throws IOException {
+    accessControlClient.checkForAccessOrThrow(
+        ResourceScope.of(accountId, null, null), Resource.of(ResourceTypes.SMTP, null), VIEW_SMTP_PERMISSION);
     NgSmtpDTO response = smtpNgService.getSmtpSettings(accountId);
     return ResponseDTO.newResponse(response);
   }
